Base implementation for Lambda
