+++ packages/lambda/config/exportServer.js
@@ -1,0 +1,5 @@
+/* eslint-disable global-require, no-console, import/order */
+const serverless = require('serverless-http');
+const app = require('@irvingjs/core/server');
+
+modules.exports.handler = serverless(app);
