+++ packages/lambda/irving.config.server.js
@@ -1,0 +1,6 @@
+const exportServer = require('./config/exportServer.js');
+
+module.exports = {
+  name: 'lambda',
+  exportServer: exportServer,
+};
